backend/requirements.txt
aio-pika==9.5.5
alembic==1.16.5
aiosqlite==0.20.0
asyncpg==0.30.0
bcrypt==4.3.0
email-validator==2.2.0
fastapi==0.116.1
httpx==0.28.1
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
pytest==8.4.2
pytest-asyncio==1.2.0
redis==6.4.0
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
